container: support re-using existing snapshots
The Layers include a digest for cached snapshots, which can be used as
the rootfs when building the layers. This patch walks through those
snapshots to identify a starting point for building containers.

Note that The actual Snapshot creation is not yet implemented.

Also, remove Delete container by name but require the caller to Get the
containe first and user the Container.Delete function.

Signed-off-by: Chris Zankel <chris@zankel.net>