container: provide commit to create the CNE container
Once a container has been built, it needs to be committed so it can be
used as an 'environment' container.

Commit updates the runspec to mount the user's home directory into the
container. It also changes the generation of the container from the base
hash used while building the container to the actual hash describing the
workspace environment.

Signed-off-by: Chris Zankel <chris@zankel.net>