suport additional envs for build execute
project:
Add a field for the additional envs to command

container:
Add envs argument to BuildExec and restructure Build to use BuildExec
instead a separate implementation

apt:
Add DEBIAN_FRONTEND=noninteractive to apt-install commands and pass it
when executing a build command.

cli/*:
Pass an empty string for the additional envs argument.

Signed-off-by: Chris Zankel <chris@zankelnet>