project: support specifying the project file name
Loading a project file required to specify the directory where the
cneporject file is located and didn't allow to specify a different
porject file name.

This patch allows callers of LoadFrom to specify a project file or
project directory. If a directory is specified, it will continue to try
to load the cneproject file

Signed-off-by: Chris Zankel <chris@zankelnet>