project: support project filenames other than 'cneproject'
Store the full project file path including the filename in the
temporary project structure. This allows users to use any file
other than 'cneproject'. This is likely more useful in combination
of using an env variable or for cloning a project.

This patch also fixes an endless loop trying to find a parent cneproject
file.

Finally, it remove NewProject and apt_test creates a skeleton Project
instance.

Signed-off-by: Chris Zankel <chris@zankelnet>