runtime: create image snapshot if it was deleted
Purging the container, for example when forcing a rebuild, deletes all
snapshots including the initial snapshot from the image. While this
should be fixed, i.e. not deleting the image snapshot, we should also
have a method to re-create that snapshot (unpacking the image), which is
implemented in this patch.

Signed-off-by: Chris Zankel <chris@zankel.net>