container: expose internal runtime container
Exposing the runtime container instance allows callers to call directoy
into the runtime interfaces instead of relaying the calls through the
container (e.g. Delete or Purge).

The goal is to have container used by the runtime implementation as it
is for local containers, so can be abstracted through the runtime.

Signed-off-by: Chris Zankel <chris@zankel.net>