cli/config: provide option to override configurations per project
Provide a configuration file override option on a per project base. The
order is:

1. system configuration
2. user coniguration
3. project configuration

The project configuration file is in the same directory as cneproject
and called cneconfig.

This also requires to make the path member in the project global to
access the cneproject path.

Finally, add cneconfig to .gitignore

Signed-off-by: Chris Zankel <chris@zankel.net>