all: remove image field in containerd.image and progress optimziations
Removing the image from the internal container struct requires to pass
it to container.Create.

Since the image information is not in the container structure anymore,
SetRootFS cannot unpack the image and create the rootfs snapshot, and
expects it to already exist.

In addition, progress output handling was broken, such as not closing
the channel from the sender, and various other synchronization issues.

Signed-off-by: Chris Zankel <chris@zankel.net>