cli/conf: add support for changing the context options
Add function to the config package to set, modify, and remove options.
Options must be provided as a comma-delimited list:
* key=value  -- to add or change an existing option
* key=       -- to remove an option

Limit the update context command to the specified configuration (user,
system, project) instead of silently failing on the merged config that
wasn't saved.

Change the signature to make UpdateContextOptions a method of the
context and remove the extra context argument.

Fix also the changes feedback: OldValue had an erraneous 1 in the name.

Signed-off-by: Chris Zankel <chris@zankel.net>