move closing the progress channel back to the CLI code
While channels should be closed by the sender, the contract for the
progress channel is that any sender won't send any more data when the
function exited. That way, the channel can be reused between
downloading, unpacking, and running commands. The channel is finally
closed when there's not furhter output.

Signed-off-by: Chris Zankel <chris@zankel.net>